Renamed telebot.types.Message.__html_text to apply_html_entities and moved it to telebot/formatting
