+++ telebot/formatting.py
@@ -4,11 +4,9 @@ Markdown & HTML formatting functions.
 .. versionadded:: 4.5.1
 """
 
-import html
-
 import re
-
-from typing import Optional
+import html
+from typing import Optional, List
 
 
 def format_text(*args, separator="\n"):
@@ -32,7 +30,6 @@ def format_text(*args, separator="\n"):
     :rtype: :obj:`str`
     """
     return separator.join(args)
-
 
 
 def escape_html(content: str) -> str:
@@ -64,6 +61,110 @@ def escape_markdown(content: str) -> str:
     parse = re.sub(r"([_*\[\]()~`>\#\+\-=|\.!\{\}\\])", r"\\\1", content)
     reparse = re.sub(r"\\\\([_*\[\]()~`>\#\+\-=|\.!\{\}\\])", r"\1", parse)
     return reparse 
+
+
+def apply_html_entities(text: str, entities: Optional[List], custom_subs: Optional[dict[str, str]]) -> str:
+    """
+    Author: @sviat9440
+    Updaters: @badiboy, @EgorKhabarov
+    Message: "*Test* parse _formatting_, [url](https://example.com), [text_mention](tg://user?id=123456) and mention @username"
+
+    .. code-block:: python3
+        :caption: Example:
+
+        apply_html_entities(text, entities)
+        >> "<b>Test</b> parse <i>formatting</i>, <a href=\"https://example.com\">url</a>, <a href=\"tg://user?id=123456\">text_mention</a> and mention @username"
+
+    Custom subs:
+        You can customize the substitutes. By default, there is no substitute for the entities: hashtag, bot_command, email. You can add or modify substitute an existing entity.
+    .. code-block:: python3
+        :caption: Example:
+
+        apply_html_entities(
+            text,
+            entities,
+            {"bold": "<strong class=\"example\">{text}</strong>", "italic": "<i class=\"example\">{text}</i>", "mention": "<a href={url}>{text}</a>"},
+        )
+        >> "<strong class=\"example\">Test</strong> parse <i class=\"example\">formatting</i>, <a href=\"https://example.com\">url</a> and <a href=\"tg://user?id=123456\">text_mention</a> and mention <a href=\"https://t.me/username\">@username</a>"
+    """
+
+    if not entities:
+        return text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;")
+
+    _subs = {
+        "bold": "<b>{text}</b>",
+        "italic": "<i>{text}</i>",
+        "pre": "<pre>{text}</pre>",
+        "code": "<code>{text}</code>",
+        # "url": "<a href=\"{url}\">{text}</a>", # @badiboy plain URLs have no text and do not need tags
+        "text_link": "<a href=\"{url}\">{text}</a>",
+        "strikethrough": "<s>{text}</s>",
+        "underline": "<u>{text}</u>",
+        "spoiler": "<span class=\"tg-spoiler\">{text}</span>",
+        "custom_emoji": "<tg-emoji emoji-id=\"{custom_emoji_id}\">{text}</tg-emoji>",
+        "blockquote": "<blockquote>{text}</blockquote>",
+    }
+
+    if custom_subs:
+        for key, value in custom_subs.items():
+            _subs[key] = value
+    utf16_text = text.encode("utf-16-le")
+    html_text = ""
+
+    def func(upd_text, subst_type=None, url=None, user=None, custom_emoji_id=None):
+        upd_text = upd_text.decode("utf-16-le")
+        if subst_type == "text_mention":
+            subst_type = "text_link"
+            url = "tg://user?id={0}".format(user.id)
+        elif subst_type == "mention":
+            url = "https://t.me/{0}".format(upd_text[1:])
+        upd_text = upd_text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;")
+        if not subst_type or not _subs.get(subst_type):
+            return upd_text
+        subs = _subs.get(subst_type)
+        if subst_type == "custom_emoji":
+            return subs.format(text=upd_text, custom_emoji_id=custom_emoji_id)
+        return subs.format(text=upd_text, url=url)
+
+    offset = 0
+    start_index = 0
+    end_index = 0
+    for entity in entities:
+        if entity.offset > offset:
+            # when the offset is not 0: for example, a __b__
+            # we need to add the text before the entity to the html_text
+            html_text += func(utf16_text[offset * 2: entity.offset * 2])
+            offset = entity.offset
+
+            new_string = func(utf16_text[offset * 2: (offset + entity.length) * 2], subst_type=entity.type,
+                              url=entity.url, user=entity.user, custom_emoji_id=entity.custom_emoji_id)
+            start_index = len(html_text)
+            html_text += new_string
+            offset += entity.length
+            end_index = len(html_text)
+        elif entity.offset == offset:
+            new_string = func(utf16_text[offset * 2: (offset + entity.length) * 2], subst_type=entity.type,
+                              url=entity.url, user=entity.user, custom_emoji_id=entity.custom_emoji_id)
+            start_index = len(html_text)
+            html_text += new_string
+            end_index = len(html_text)
+            offset += entity.length
+        else:
+            # Here we are processing nested entities.
+            # We shouldn't update offset, because they are the same as entity before.
+            # And, here we are replacing previous string with a new html-rendered text(previous string is already html-rendered,
+            # And we don't change it).
+            entity_string = html_text[start_index: end_index].encode("utf-16-le")
+            formatted_string = func(entity_string, subst_type=entity.type, url=entity.url, user=entity.user,
+                                    custom_emoji_id=entity.custom_emoji_id). \
+                replace("&amp;", "&").replace("&lt;", "<").replace("&gt;", ">")
+            html_text = html_text[:start_index] + formatted_string + html_text[end_index:]
+            end_index = len(html_text)
+
+    if offset * 2 < len(utf16_text):
+        html_text += func(utf16_text[offset * 2:])
+
+    return html_text
 
 
 def mbold(content: str, escape: Optional[bool]=True) -> str:
